+++ src/vector/distance.cpp
@@ -1,13 +1,13 @@
-#include "vector/distance.h"
+#include "vector/distance.hpp"
 
-#include "db/errors.h"
+#include "db/errors.hpp"
 
 #include <cmath>
 
 namespace vrdb {
 namespace {
 
-void validateSameDimension(const std::vector<float>& a, const std::vector<float>& b) {
+void validateSameDimension(const std::vector<double>& a, const std::vector<double>& b) {
     if (a.size() != b.size()) {
         throw QueryError("vectors must have matching dimensions");
     }
@@ -15,37 +15,37 @@ void validateSameDimension(const std::vector<float>& a, const std::vector<float>
 
 } // namespace
 
-float cosineDistance(const std::vector<float>& a, const std::vector<float>& b) {
+double cosineDistance(const std::vector<double>& a, const std::vector<double>& b) {
     validateSameDimension(a, b);
 
-    float dot = 0.0f;
-    float normA = 0.0f;
-    float normB = 0.0f;
+    double dot = 0.0;
+    double normA = 0.0;
+    double normB = 0.0;
     for (std::size_t i = 0; i < a.size(); ++i) {
         dot += a[i] * b[i];
         normA += a[i] * a[i];
         normB += b[i] * b[i];
     }
 
-    if (normA == 0.0f || normB == 0.0f) {
+    if (normA == 0.0 || normB == 0.0) {
         throw QueryError("cosine distance is undefined for zero vectors");
     }
 
-    return 1.0f - dot / (std::sqrt(normA) * std::sqrt(normB));
+    return 1.0 - dot / (std::sqrt(normA) * std::sqrt(normB));
 }
 
-float euclideanDistance(const std::vector<float>& a, const std::vector<float>& b) {
+double euclideanDistance(const std::vector<double>& a, const std::vector<double>& b) {
     validateSameDimension(a, b);
 
-    float sum = 0.0f;
+    double sum = 0.0;
     for (std::size_t i = 0; i < a.size(); ++i) {
-        const float delta = a[i] - b[i];
+        const double delta = a[i] - b[i];
         sum += delta * delta;
     }
     return std::sqrt(sum);
 }
 
-float distance(const std::vector<float>& a, const std::vector<float>& b, DistanceMetric metric) {
+double distance(const std::vector<double>& a, const std::vector<double>& b, DistanceMetric metric) {
     switch (metric) {
     case DistanceMetric::EUCLIDEAN:
         return euclideanDistance(a, b);
